+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * New cop `BlockNesting` checks for excessive block nesting
 * Added `-f`/`--format` option to specify formatter.
+* Added `-o`/`--out` option to specify output file for each formatter.
 * Added `--require` option to inject external Ruby code into RuboCop.
 
 ### Changes